chore(be): change business logic column and template for 3 columns default

+++ backend/task-service/src/main/java/com/ntt/task_service/service/ColumnService.java
@@ -91,6 +91,8 @@ public class ColumnService {
 
         projectAuthorizationService.validateCanManage(projectId);
 
+        if (Boolean.TRUE.equals(column.getIsDoneColumn())) throw new AppException(ErrorCode.CANNOT_DELETE_DONE_COLUMN);
+
         columnRepository.delete(column);
     }
